Moved charts to the dashboard page
